+++ test2.py
@@ -1,0 +1,5 @@
+def add(numbers) -> int:
+    res = 0
+    for n in numbers:
+        res += n
+    return res